api: use ArrayVec for directions in StopInfoRoutes
There is a rare case where there's only one direction present in the data. This can result in a parsing error in Serde deserialization. This field shouldn't surpass 2 elements, so using a Vec here seems unnecessary. Using an ArrayVec here can convey the business logic in the code and avoid allocation on the heap.